relocate http query params stuff to util package
